# Apply micro-optimisation for Route API